requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
django-bower==5.1.0
django-crispy-forms==1.6.0
django-debug-toolbar==1.5
django-extensions==1.6.7
django-material==0.8.0
django==1.9.7
pdfminer3k==1.3.0
ply==3.9
py==1.4.31
pytest==3.0.2
six==1.10.0
sqlparse==0.2.0
uwsgi==2.0.13.1
